requirements.txt

matplotlib==3.5.2
streamlit==1.10.0
pandas==1.4.2
Pillow==9.1.1

